Rename directory with page objects
