+++ e2e/about_page_spec.js
@@ -1,4 +1,4 @@
-const Page = require('./support/page');
+const Page = require('./page_objects/page');
 
 class AboutPage extends Page {
 
@@ -9,7 +9,7 @@ describe('About page', () => {
   const page = new AboutPage('/#/about');
 
   beforeEach(() => {
-    page.open();
+    page.navigate();
   });
 
   it('has a title', () => {
