+++ e2e/contacts_page_spec.js
@@ -1,6 +1,6 @@
-const Page = require('./support/page');
-const Table = require('./support/table');
-const FavouriteButton = require('./support/favourite_button');
+const Page = require('./page_objects/page');
+const Table = require('./page_objects/table');
+const FavouriteButton = require('./page_objects/favourite_button');
 
 class ListPage extends Page {
 
@@ -24,7 +24,7 @@ describe('Contacts page', () => {
   const page = new ListPage('/#/contacts/');
 
   beforeEach(() => {
-    page.open();
+    page.navigate();
   });
 
   describe('list', () => {
