+++ e2e/page_objects/favourite_button.js
